Add curse version parsing logic
In order to facilitate the ability to follow alpha/beta releases of
addons, we need to source the latest releases from the /files endpoint
of a given addon on curseforge. The `versions()` generator and
`CurseAddonVersion` object provide abstractions for iterating through
the list of all releases alpha/beta/release for any given GameVersion.

Note, the `game_version_filter` string is used in the URL params to
effectively select from the dropdown the appropriate game version
when searching for the release list.

Currently `get_latest_version` is broken due to searching for the
wrong version string.