Add logic to track alpha/beta/release versions
Includes the `AddonVersion` enum to provide an ordering
of release precedent and the `Curse` object takes an element of
this enum when being instantiated. Defaults to tracking "release"
versions of the addon.

The important logic falls in the `get_latest_version` function where
it selects the addon the meets or exceeds the desired track. Tracking
`AddonVersion.alpha` will always grab the latest version, whereas
`AddonVersion.beta` will ignore alpha releases and so on.